adding some generated code from axis-2
This code does not work at all, and it looks
like shit. This is all auto generated from axis-2
using the WSDL from SSO. LOTS of work needs to be
done but I figured Id check this in for now as I
work on it. This is the beginning of issue #6