Adds setAlarmStatus
This commit adds the ability to reset an alarm from red to green
using a hidden API call. This call should work similar to the
acknowledgeAlarm() call but requires an additional param of the
alarm color. To reset from red to green the alarm status passed
in should be that of "green".