Fixes #28
This commits fixes the only throwing RemoteException issue
raised in #28 There still needs to be more testing added
but this commit for sure addresses the InvalidLogin fault
not being thrown, and looks like it will fix others.

Further refactoring will be needed to be able to fully unit
test the Clients though.