Changed Long[] and Integer[] to correct types
When an optional property is found and is of type long or
int it needed to be Long or Integer. It is not the same for
an optional array of those types. When it is an array of int
or long it should be those basic types. This was fixed in
yavijava_generator/19

Fixes #93