Refactor TrustAllSSL and WSClient to fix security
The way the library has always worked the default ssl
context (when using ignore certs) would change the global
context for anything in the jvm. This means ALL certs were
trusted when using ignoreCert in YAVIJAVA/VIJAVA. This is
a big security issue. A user on the vijava forums pointed
the issue out: http://sourceforge.net/p/vijava/discussion/826592/thread/33f68590/
This fixes that issue and includes tests to verify fix.

Fixes #38