Fixes issue in ReflectUtil with String[]
In the ReflectUtil.parseToObject method if a String[]
was being returned it would actually return a List
which would cause a class cast exception later when the
object was being accessed. This change set includes both
unit tests for the XML parser with 2 types of String Array
payloads and also a test on the ReflectUtil.parseToObject
method.

Fixes #186